move all autos to 'autos' namespace
