+++ src/client_handler.cpp
@@ -52,7 +52,7 @@ void ClientHandler::handle_get_existing_rules(ClientMessage &client_message)
 {
     ExistingRulesServerMessageContent content;
     {
-        AutoLock auto_rules_lock(_rules_manager->get_lock());
+        autos::AutoLock auto_rules_lock(_rules_manager->get_lock());
         for (auto &rule : *_rules_manager)
         {
             content.add_rules()->CopyFrom(rule);
