+++ src/common/autos/auto_temp_folder.cpp
@@ -1,7 +1,7 @@
 #include "auto_temp_folder.hpp"
 #include "../os_utils.hpp"
 
-os::AutoTempFolder::AutoTempFolder(const std::string &template_name)
+autos::AutoTempFolder::AutoTempFolder(const std::string &template_name)
     : _path(template_name)
 {
     if (MOUNT_ERROR_VALUE == mkdtemp(const_cast<char *>(_path.c_str())))
@@ -10,12 +10,12 @@ os::AutoTempFolder::AutoTempFolder(const std::string &template_name)
     }
 }
 
-os::AutoTempFolder::~AutoTempFolder()
+autos::AutoTempFolder::~AutoTempFolder()
 {
     std::filesystem::remove_all(_path);
 }
 
-std::filesystem::path os::AutoTempFolder::get() const
+std::filesystem::path autos::AutoTempFolder::get() const
 {
     return _path;
 }
